Criado etapa de home

+++ lib/ui/screens/publication_details/adoption_post_details_screen.dart
@@ -86,8 +86,10 @@ class _AdoptionDetailsScreenState extends State<AdoptionDetailsScreen>
     DocumentSnapshot<Map<String, dynamic>> dataUser =
         await userService.getUserProfile(_idUser.value!);
     if (dataUser.data() != null) {
-      _listFavoritesFirebase =
-          List<String>.from(dataUser.data()?['listFavoritesAnimal']);
+      if (dataUser.data()?['listFavoritesAnimal'] != null) {
+        _listFavoritesFirebase =
+            List<String>.from(dataUser.data()?['listFavoritesAnimal']);
+      }
       _isSelected = _listFavoritesFirebase.contains(_idPublication.value);
       latUser = dataUser.data()?['lat'] ?? 0.0;
       longUser = dataUser.data()?['long'] ?? 0.0;
@@ -184,7 +186,7 @@ class _AdoptionDetailsScreenState extends State<AdoptionDetailsScreen>
   Widget build(BuildContext context) {
     final animalModel = context.read<PublicationModel>();
     return Scaffold(
-      floatingActionButton: status != 'finished'
+      floatingActionButton: status != 'finished' && _isMyPublication
           ? SizedBox(
               height: 70,
               width: 70,
@@ -430,40 +432,42 @@ class _AdoptionDetailsScreenState extends State<AdoptionDetailsScreen>
                                   ),
                               ],
                             ),
-                            const SizedBox(height: 64),
-                            ButtonComponent(
-                              text: 'Finalizar publicação',
-                              color: const Color(0xff21725E),
-                              onPressed: () {
-                                Navigator.pushNamed(
-                                    context, '/end_publication');
-                              },
-                            ),
-                            const SizedBox(height: 16),
-                            ButtonComponent(
-                              text: 'Excluir publicação',
-                              color: const Color(0xffA82525),
-                              onPressed: () async {
-                                await showDialog(
-                                  context: context,
-                                  builder: (context) =>
-                                      const AlertDialogComponent(
-                                    statusType: 'error',
-                                    title: 'Excluir publicação',
-                                    message:
-                                        'A publicação será excluída permanentemente. Deseja prosseguir ?',
-                                  ),
-                                ).then((value) {
-                                  if (value) {
-                                    PublicationService.deletePublication(
-                                        _idPublication.value!,
-                                        "publications_animal");
-                                    Navigator.pushReplacementNamed(
-                                        context, '/my_publications');
-                                  }
-                                });
-                              },
-                            ),
+                            if (_isMyPublication) ...[
+                              const SizedBox(height: 64),
+                              ButtonComponent(
+                                text: 'Finalizar publicação',
+                                color: const Color(0xff21725E),
+                                onPressed: () {
+                                  Navigator.pushNamed(
+                                      context, '/end_publication');
+                                },
+                              ),
+                              const SizedBox(height: 16),
+                              ButtonComponent(
+                                text: 'Excluir publicação',
+                                color: const Color(0xffA82525),
+                                onPressed: () async {
+                                  await showDialog(
+                                    context: context,
+                                    builder: (context) =>
+                                        const AlertDialogComponent(
+                                      statusType: 'error',
+                                      title: 'Excluir publicação',
+                                      message:
+                                          'A publicação será excluída permanentemente. Deseja prosseguir ?',
+                                    ),
+                                  ).then((value) {
+                                    if (value) {
+                                      PublicationService.deletePublication(
+                                          _idPublication.value!,
+                                          "publications_animal");
+                                      Navigator.pushReplacementNamed(
+                                          context, '/my_publications');
+                                    }
+                                  });
+                                },
+                              ),
+                            ],
                           ] else
                             Container(
                               padding: const EdgeInsets.all(16),
